backend/requirements.txt
azure-ai-documentintelligence>=1.0.0
azure-core>=1.30.0
python-dotenv>=1.0.0

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9